Remote Connections API (#64)
